Fix: Use timezone-aware datetime for schedule date calculation
The bug occurred when checking the game schedule at 7:00 PM EST - the date would
suddenly switch from Nov 20 to Nov 21 because datetime.now() returns UTC time
(which is 00:00 UTC = midnight on the next day when it's 19:00 EST).

This caused the system to look for games on the wrong date and enter a bad state
where it thought a game was 'about to start' but couldn't find it in the schedule.

Changed from: datetime.now().strftime('%Y-%m-%d')
Changed to: datetime.now(pytz.timezone(TIMEZONE)).strftime('%Y-%m-%d')

This ensures the schedule check uses Eastern Time (America/New_York) consistently
with the rest of the application logic.